Fix 422 export error: ShapeOverlay stroke_width rejects float values
The shape overlay on the timeline has stroke_width=13.5 but the backend
model defined it as int, causing Pydantic to reject it. Changed
stroke_width and corner_radius to float to accept fractional values.

+++ backend/models.py
@@ -221,8 +221,8 @@ class ShapeOverlay(BaseModel):
     height: float = 20         # size % (0-100)
     fill_color: str = "#FF3B30"
     stroke_color: str = "#FFFFFF"
-    stroke_width: int = 2
-    corner_radius: int = 0     # for rectangles
+    stroke_width: float = 2
+    corner_radius: float = 0   # for rectangles
     start_time: float = 0.0
     end_time: float = 0.0
     rotation: float = 0.0
